Move custom exceptions, refactor exceptionHandler
