+++ src/main/java/com/shape/shape_api/exception/InvalidEntityException.java
@@ -1,4 +1,4 @@
-package com.shape.shape_api.error;
+package com.shape.shape_api.exception;
 
 public class InvalidEntityException extends RuntimeException {
     public InvalidEntityException(String message) {
