+++ src/main/java/com/shape/shape_api/exception/InvalidShapeParameterValueException.java
@@ -1,0 +1,7 @@
+package com.shape.shape_api.exception;
+
+public class InvalidShapeParameterValueException extends RuntimeException {
+    public InvalidShapeParameterValueException(String message) {
+        super(message);
+    }
+}
