+++ src/main/java/com/shape/shape_api/exception/MissingParameterException.java
@@ -1,4 +1,4 @@
-package com.shape.shape_api.error;
+package com.shape.shape_api.exception;
 
 public class MissingParameterException extends RuntimeException {
     public MissingParameterException(String message) {
